Fixed issues with Bow, including gameplay changes.
Added Charging Movescale to table to make charging attacks disallow
movement. Adjusted rotation in Fighter to allow facing in movedirection
even with a movespeed of 0, so that an untargeted heavy bow attack
can rotate in place. Fixed Damage of projectiles to scale with
charge percentage. Fixed bug with AttackCast so that projectiles
don't collide with their parents.